made back to story button reusable
