+++ client/src/components/GoToStoryButton/GoToStoryButton.scss
@@ -1,0 +1,5 @@
+.gotostory-btn {
+    position: absolute;
+    top: 120px;
+    left: 50px;
+}
