+++ client/src/controllers/Entry/Entry.scss
@@ -6,12 +6,6 @@
     margin: 300px auto;
 }
 
-.gotostory-btn {
-    position: absolute;
-    top: 120px;
-    left: 50px;
-}
-
 .entry-container {
     display: flex;
     flex-direction: column;
