+++ client/src/controllers/EntryEdit/EntryEdit.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import './EntryEdit.scss'
+
+const EntryEdit = props => {
+    return (
+
+    )
+}
+
+export default EntryEdit
