+++ client/src/controllers/Story/Story.js
@@ -86,7 +86,7 @@ const Story = props => {
                 <StoryCard
                     {...{
                         story: story,
-                        specialStyle:{margin: '0px auto 50px auto', border:'none', cursor:'auto', opacity:1},
+                        specialStyle:{margin: '0px auto 50px auto', cursor:'auto', opacity:1},
                         inTimeline: true
                     }}
                 />
