+++ client/src/controllers/StoryCreate/StoryCreate.scss
